refactor(routes): extract page data hook
Move the Routes page data orchestration into useRoutesData so the page can focus on layout and rendering.\n\nThe new hook now owns route-query consumption, search derivation, recently viewed route handling, message selection, and list scroll reset behavior without changing user-visible results.\n\nCo-authored-by: Codex gpt-5.4 <noreply@openai.com>
